Refactor MediTwin Backend
