+++ backend/.env
@@ -1,0 +1,9 @@
+# Gemini API Configuration
+GEMINI_API_KEY=your_gemini_api_key
+
+# OpenFDA API Configuration
+OPENFDA_API_KEY=your_openfda_api_key
+
+# Flask Configuration
+SECRET_KEY=your_secret_key
+FLASK_DEBUG=True
