+++ backend/config.py
@@ -1,0 +1,16 @@
+import os
+from dotenv import load_dotenv
+
+# Load environment variables from .env file
+load_dotenv()
+
+class Config:
+    # Gemini API Configuration
+    GEMINI_API_KEY = os.getenv('GEMINI_API_KEY', '')
+    
+    # OpenFDA API Configuration
+    OPENFDA_API_KEY = os.getenv('OPENFDA_API_KEY', '')
+    
+    # Flask Configuration
+    SECRET_KEY = os.getenv('SECRET_KEY', 'default-secret-key')
+    DEBUG = os.getenv('FLASK_DEBUG', 'False') == 'True'
