+++ backend/requirements.txt
@@ -1,0 +1,6 @@
+flask
+flask-cors
+requests
+python-dotenv
+google-generativeai
+gunicorn
